rbd: allocate memory for rbd_snap_list
This fixes RbdImage.snapList by allocating memory for librbd to fill in.

Note that this is a temporary fix. I think that the setAutoSynch(false)
trick is a hack, and should be replaced with something more elegant,
although my JNA-fu is not excellent, so it's not immediately clear what
it should look like.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>